refactor batch export
